chore: build @qvac/logging from TypeScript sources
