+++ packages/logging/.lunteignore
@@ -1,2 +1,2 @@
-# FIXME Remove this once the Lunte bug is fixed.
-index.d.ts
+dist
+test/dist
