--- packages/logging/env.js
@@ -1,1 +1,0 @@
-module.exports = typeof process === 'undefined' ? {} : process.env
